+++ mark1/lib/configs/themes/theme_provider.dart
@@ -1,12 +1,11 @@
-// ignore_for_file: avoid_print
-
 import 'package:flutter/material.dart';
+import 'package:mymy_m1/helpers/logs/log_helper.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:mymy_m1/configs/themes/theme_collections.dart';
 
 class ThemeProvider with ChangeNotifier {
   ThemeMode _themeMode = ThemeMode.system;
-  String _currentThemeName = 'sixPM';
+  String _currentThemeName = 'original';
   late SharedPreferences _prefs;
 
   ThemeProvider() {
@@ -43,7 +42,7 @@ class ThemeProvider with ChangeNotifier {
     _prefs = await SharedPreferences.getInstance();
     _themeMode =
         ThemeMode.values[_prefs.getInt('themeMode') ?? ThemeMode.system.index];
-    _currentThemeName = _prefs.getString('themeName') ?? 'sixPM';
+    _currentThemeName = _prefs.getString('themeName') ?? 'original';
 
     // Apply user preference before falling back to system default
     if (_themeMode == ThemeMode.system) {
@@ -59,7 +58,7 @@ class ThemeProvider with ChangeNotifier {
   Future<void> _saveToPrefs() async {
     await _prefs.setInt('themeMode', _themeMode.index);
     await _prefs.setString('themeName', _currentThemeName);
-    print(
+    LogHelper.logger.i(
         "\t New Prefs: Theme name: $_currentThemeName | Theme mode: ${ThemeMode.values[_prefs.getInt('themeMode') ?? ThemeMode.system.index].toString().split('.').last} ");
   }
 }