+++ mark1/lib/helpers/templates/main_view_template.dart
@@ -1,11 +1,19 @@
 import 'package:flutter/material.dart';
 
 Scaffold mainView(BuildContext context,
-    {String? appBarTitle = 'changeMe', required Widget body}) {
+    {String? appBarTitle = 'changeMe',
+    required Widget body,
+    List<Widget>? appbarActions = const [],
+    Color? appBarBackgroundColor}) {
   return Scaffold(
       appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(appBarTitle!),
+        backgroundColor: appBarBackgroundColor ??
+            Theme.of(context).colorScheme.onPrimaryFixed,
+        title: Text(
+          appBarTitle!,
+          style: TextStyle(color: Theme.of(context).colorScheme.primary),
+        ),
+        actions: appbarActions,
       ),
       backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
       body: body);