+++ mark1/lib/main.dart
@@ -1,20 +1,24 @@
-// ignore_for_file: avoid_print
-
+import 'package:loading_animation_widget/loading_animation_widget.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:logger/logger.dart';
 import 'package:mymy_m1/configs/languages/language_provider.dart';
 import 'package:mymy_m1/configs/themes/theme_provider.dart';
 import 'package:mymy_m1/configs/themes/theme_collections.dart';
+import 'package:mymy_m1/firebase_options.dart';
+import 'package:mymy_m1/helpers/logs/log_helper.dart';
 import 'package:mymy_m1/navigation/pages_router.dart';
-import 'package:mymy_m1/services/notifications/custom_notification_listener.dart';
-import 'package:mymy_m1/services/notifications/custom_notification_service.dart';
+import 'package:mymy_m1/helpers/getit/get_it.dart';
+import 'package:mymy_m1/services/notifications/notification_manager.dart';
 import 'package:mymy_m1/services/settings/settings_service.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() async {
+  LogHelper.logger.d("Logger is working!");
   WidgetsBinding widgetsBinding = WidgetsFlutterBinding.ensureInitialized();
   FlutterNativeSplash.preserve(widgetsBinding: widgetsBinding);
 
@@ -29,46 +33,52 @@ void main() async {
     languageProvider: languageProvider,
   );
 
-  // Print user preferences
+  // Log user preferences
   SharedPreferences prefs = await SharedPreferences.getInstance();
-  print("\n<------------------->");
-  print(' User Preferences:');
-  print('Language Code: ${prefs.getString('languageCode')}');
-  print('Is System Default Language: ${prefs.getBool('isSystemDefault')}');
-  print(
+  // LogHelper.logger.d("\n<------------------->");
+  LogHelper.logger.i(' User Preferences:');
+  LogHelper.logger.i('Language Code: ${prefs.getString('languageCode')}');
+  LogHelper.logger
+      .d('Is System Default Language: ${prefs.getBool('isSystemDefault')}');
+  LogHelper.logger.i(
       'Theme Mode: ${ThemeMode.values[prefs.getInt('themeMode') ?? ThemeMode.system.index].toString().split('.').last}');
-  print('Theme Name: ${prefs.getString('themeName')}');
-  print("<------------------->");
+  LogHelper.logger.i('Theme Name: ${prefs.getString('themeName')}');
+  // LogHelper.logger.d("<------------------->");
 
-  // Print what will be applied
-  print("\n<------------------->");
-  print(' Applied Settings:');
-  print(
+  // Log what will be applied
+  // LogHelper.logger.d("\n<------------------->");
+  LogHelper.logger.i(' Applied Settings:');
+  LogHelper.logger.i(
       'Language: ${languageProvider.currentLocale.languageCode} (Is System Default: ${languageProvider.isSystemDefault})');
-  print('Theme Mode: ${themeProvider.themeMode.toString().split(".").last}');
-  print('Theme Name: ${themeProvider.currentThemeName}');
-  print("<------------------->");
+  LogHelper.logger
+      .d('Theme Mode: ${themeProvider.themeMode.toString().split(".").last}');
+  LogHelper.logger.i('Theme Name: ${themeProvider.currentThemeName}');
+  // LogHelper.logger.d("<------------------->");
 
   // Report System default
-  print("\n<------------------->");
-  print(" Device Default:");
+  // LogHelper.logger.d("\n<------------------->");
+  LogHelper.logger.i(" Device Default:");
   String deviceLocale =
       WidgetsBinding.instance.platformDispatcher.locales.toString();
-  print('Device Default Locale: $deviceLocale');
-  print(
+  LogHelper.logger.i('Device Default Locale: $deviceLocale');
+  LogHelper.logger.i(
       "Device Default Theme mode: ${WidgetsBinding.instance.platformDispatcher.platformBrightness.toString().split(".").last}");
-  print("<------------------->");
+  // LogHelper.logger.d("<------------------->");
 
-  // Delay to allow time for reading the console output
-  await Future.delayed(const Duration(seconds: 1));
+  await Future.delayed(const Duration(
+      seconds: 1)); // Delay to allow time for reading the console output
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
 
   runApp(
     MultiProvider(
       providers: [
-        ChangeNotifierProvider.value(value: themeProvider),
-        ChangeNotifierProvider.value(value: languageProvider),
+        // ChangeNotifierProvider.value(value: themeProvider),
+        // ChangeNotifierProvider.value(value: languageProvider),
         ChangeNotifierProvider.value(value: settingsService),
-        ChangeNotifierProvider(create: (_) => CustomNotificationService()),
+        ProxyProvider<SettingsService, NotificationManager>(
+          update: (_, settings, __) => NotificationManager(settings),
+          dispose: (_, manager) => manager.dispose(),
+        ),
       ],
       child: const MyApp(),
     ),
@@ -83,8 +93,6 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
-  final CustomNotificationService _notificationService =
-      CustomNotificationService();
   late bool _isLoading;
 
   @override
@@ -97,29 +105,31 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
 
   void initialization() async {
     setState(() => _isLoading = true);
-    print("\n<------------------->");
-    print(" Available Resources:");
-    print("loading resources...");
-    print("Available localizations: ${AppLocalizations.supportedLocales}");
-    print("Available Themes: ${ThemeCollections.availableThemes}");
+    // LogHelper.logger.d("\n<------------------->");
+    LogHelper.logger.i(" Available Resources:");
+    LogHelper.logger.i("loading resources...");
+    setupDependencies();
+    LogHelper.logger
+        .d("Available localizations: ${AppLocalizations.supportedLocales}");
+    LogHelper.logger.i("Available Themes: ${ThemeCollections.availableThemes}");
     // await Future.delayed(const Duration(seconds: 1));
     setState(() => _isLoading = false);
-    print("resources are loaded successfully\n");
-    print("<------------------->\n");
+    LogHelper.logger.t("resources are loaded successfully\n");
+    // LogHelper.logger.d("<------------------->\n");
     FlutterNativeSplash.remove();
   }
 
   @override
   void dispose() {
     WidgetsBinding.instance.removeObserver(this);
-    _notificationService.dispose();
+    // TODO: add dispose()
     super.dispose();
   }
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
     if (state == AppLifecycleState.detached) {
-      _notificationService.dispose();
+      // TODO: add dispose()
     }
   }
 
@@ -129,30 +139,30 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
     return Consumer<SettingsService>(
       builder: (context, settings, child) {
         return _isLoading
-            ? const Center(
-                child: CircularProgressIndicator(),
+            ? Center(
+                child: LoadingAnimationWidget.twoRotatingArc(
+                    color: Theme.of(context).colorScheme.onSurface, size: 20),
               )
-            : MaterialApp.router(
-                routerConfig: router,
-                localizationsDelegates: const [
-                  AppLocalizations.delegate,
-                  GlobalMaterialLocalizations.delegate,
-                  GlobalWidgetsLocalizations.delegate,
-                  GlobalCupertinoLocalizations.delegate,
-                ],
-                supportedLocales: settings.supportedLocales,
-                locale: settings.currentLocale,
-                title: "myMY M1 by LukeCreated",
-                theme: settings.lightTheme,
-                darkTheme: settings.darkTheme,
-                themeMode: settings.themeMode,
-                builder: (context, child) {
-                  return ScaffoldMessenger(
-                    child: CustomNotificationListener(
-                      child: child ?? const SizedBox.shrink(),
-                    ),
-                  );
-                });
+            : SafeArea(
+                child: MaterialApp.router(
+                    debugShowCheckedModeBanner: false,
+                    routerConfig: router,
+                    localizationsDelegates: const [
+                      AppLocalizations.delegate,
+                      GlobalMaterialLocalizations.delegate,
+                      GlobalWidgetsLocalizations.delegate,
+                      GlobalCupertinoLocalizations.delegate,
+                    ],
+                    supportedLocales: settings.supportedLocales,
+                    locale: settings.currentLocale,
+                    title: "myMY M1 by LukeCreated",
+                    theme: settings.lightTheme,
+                    darkTheme: settings.darkTheme,
+                    themeMode: settings.themeMode,
+                    builder: (context, child) {
+                      return child ?? const SizedBox.shrink();
+                    }),
+              );
       },
     );
   }